requirements.txt
hdf5storage==0.1.19
tqdm
scipy==1.10.0
numpy==1.26.4
# matplotlib==3.9.2
torch==2.2.2
torchvision==0.17.2
# opencv-python==4.10.0.84
Pillow~=11.0.0
# pytesseract==0.3.13
#tesseract #==4.1.0